build: validate --with-formatid input
To ensure validity of the format number hardcoded in the binary, check
the supplied input against a regular expression using grep.

Note that the double square brackets are needed for the regex character
classes because square brackets are used for M4 quotation in Autoconf;
an extra pair of square brackets is needed to prevent them from being
stripped in the configure script.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>