Add inconsistent version files check while re-updating
Adds a check to compare version numbers in <state_dir>/version and
<target_dir>/usr/lib/os-release when attempting to re-update after a
format bump.

os-release should always be equal to <state_dir>/version for normal
updates and potentially greater than <state_dir>/version for format
bumps.

In the event of a bad build where os-release does not get updated to the
new version, this check will prevent an infinite loop where swupd reads
the current version from os-release, sees it is out of date, attempts to
update again, but the file does not get updated.

This occasion is very unlikely but the guard is nice to have
re(guard)less.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>