Add basic progress status for long-running tasks
Add a basic progress indication for a couple potentially long-running
swupd tasks (download_pack and add_missing_files). The progress
indication is displayed as a percentage:

...85%

It overwrites its own line so the progress will update in place. It may
be interrupted by other swupd output:

...45%
Extracting os-core pack for version 10000
...50%
Extracting os-core-update pack for version 9000
...55%

The progress indication will then resume updating in place, on the same
line.

This progress is only displayed when outputting to a TTY, and will not
output anything when the output is being redirected. This is nice for
testing and logging, but means that these progress updates will not be
caught by processes that capture swupd output, such as ister.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>