Add progress updates for redirected output
Add a progress indication for redirected output such as for logging,
testing, or third-party software such as ister.

For every 10 files or steps completed a dot (".") is printed to the
screen.

Adds a few more progress indications for silent loops.

Also adds leading newlines to some existing print statements so they are
printed on their own line.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>