Add bundlelist --deps subcommand
Allow user to list all bundle dependencies of a bundle passed to

swupd bundle-list --deps BUNDLE

This lists all included bundles (included those recursively included)
for the BUNDLE. This is particularly useful when applications or users
need to determine just how large the bundle addition will be. For
example, a user may not be aware that by bundle-adding desktop-dev they
are recursively adding 12 other bundles. This command allows the user to
make this check before installation.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>