Special handling for pack download/extract progress
Printing a percentage status for pack download and extraction is
somewhat meaningless, since the list will not be very long and output is
already sufficiently verbose.

Instead, update the pack fprintfs to include the current and total
number of packs. Print the pack name for every pack, not just those that
are being extracted so that the "(pack n of N)" message is meaningful.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>