Improve error reporting for bundleadd hash check
Add a check for missing file so that bundleadd can fall back to the
verify_fix_path to attempt to re-download the file instead of failing
the hash check due to the hash being all 0s.

Improve error messages by reporting the filename instead of the hash in
the error.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>