Fix renames of files without changes
Fixes #272

The create_update_list function was not adding a file onto its return
list if it was part of a rename that didn't change the contents. This
was due to a short circuit which failed to check if the flags
associated with a file (is_file, is_deleted etc) were the same.

Signed-off-by: Icarus Sparry <icarus.w.sparry@intel.com>