Update fix-missing-file test to account for progress dots
Now we have dots (or percentages) when files are being downloaded
these can intermingle with the normal output if they are flushed.

Alter the fix-missing-file test to remove dots from the lines to be
compared, and alter the lines-checked to add in the lines which would
normally be ignored but no longer match due to the missing dots,
e.g. "Downloading packs."

Signed-off-by: Icarus Sparry <icarus.w.sparry@intel.com>