Fix progress meters
The denominator for the progress report is based on the length of the
linked list to be processed, and the numerator is the number of
elements processed. Most loops have short cut conditions, which invoke
continue, which don't therefore update the numerator, and so loops end
before the progress meter shows 100%. In particular the final file
might be skipped.

Update the numerator early in the loop, and add a final call to
print_progress to force showing 100% if it has not already been done.

Note this can force another dot to be output for the case where stdout
is not connected to a terminal.

Signed-off-by: Icarus Sparry <icarus.w.sparry@intel.com>