downloads: support mixing http and file URLs
swupd used to enforce that version and content URL had to use the same
protocol because then it only needed to check once whether it was
downloading locally.

However, it's not that much more difficult to check each individual
URL, so now that's what swupd does now and thus the restriction is
gone.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>