Fix incorrect cycle.
Remove the increment left-over from changing from while to for.

The issue does not impact happy scenario. However, when the fallback
needs to be used (e.g. when the default trust store is removed or in a
container where it's not generated), swupd fails to connect because it
never gets to try the fallback CApath: if there's just one fallback, it
the cycle will end before it is tried.