test: add lines-checked for update/rename test
Since I changed check_lines() to propogate the matcher.awk exit status,
this test failed due to a missing 'lines-checked' file. Add the file to
fix that issue.

Also, remove the -x and -v options to the shell invocation, because more
lines of output are produced and that should not be checked.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>