Implement ghosted file handling
Implement handling of ghosted files in manifests via the 'g' flag. These
files are treated the same as deleted files except that they are not
actually deleted during the update. This allows ghosted files to serve
as the basis for renames while not actually being deleted by swupd.

This change also covers the verify --fix and verify --fix --picky paths.
Ghosted files will be ignored in these cases.

An example of where this is useful is with kernel files, which are
deleted from /usr/lib/kernel/ by clr-boot-manager but should still be
marked as deleted to utilize the rename functionality.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>