Do not fail on verifytime failure
Fixes #234
Fixes #289

The verifytime program is a helper program used by swupd to correct
certain time skews. If verifytime fails, however, this should not be a
failure for swupd. Swupd should continue to try and verify the signature
if possible. Instead of failing print out a warning to give the user a
hint to why their signature verification may fail.

The use case this solves is when the /usr/share/clear/versionstamp file
does not exist, verifytime will fail even if the system time is in an OK
state.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>