Release v3.13.0
This release adds mixer integration to swupd-client which allows local
content to be added alongside upstream content.

A bug is also fixed where CTRL-C during fullfile download caused
subsequent operations to fail due to a hash mismatch in the state
directory. Autocompletion was fixed. A --no-scripts flag was added to
the post-update scripts entirely, mainly for container use-cases when
these scripts will not work.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>