Do not try to resume failed bundle-add pack downloads
Fixes #352
This operation is fragile and can cause errors if a zero pack download
is interrupted resulting in an incomplete tar archive. A user reported
that having the incomplete pack around prevented the pack from being
re-downloaded and the content was incorrect so the bundle-add was not
completing. Instead of trying to resume a bundle-add pack download just
remove the old file and try again.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>