Release v3.14.4
This release calls clr-service-restart if a POST_UPDATE script is not
configured and the clr-service-restart binary exists on the system.
clr-service-restart is a program that restarts all services that are
running against out-of-date service files or libraries. Any POST_UPDATE
script should also call this binary in a similar fashion.

This release also fixes a couple memory leaks and a warning call in
archive.c libarchive calling code.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>