Properly initalize/deinit all bundle list variants
Let the bundle_list_main call init and deinit, don't exit()
early. Also, simplify the conditions dispatching the individual
commands. This fixes as leaks in various cases and double
initialization listing installable bundles.

Calling exit directly was used for a particular case, before init was
called, to prevent curl to be initialized for a case that didn't need
it. Then the code moved but the exit() call remained and got used by
other cases.