Fix "Hardlink target error" when extracting
In some cases packs may include files with entries of type "hard
link", that refer to previous files in the pack. Libarchive was
failing since it couldn't find the previous file in disk to archive,
because we change its output path.

Adjusting the hardlink target path (when it exists) fix the
issue. This is also what bsdtar does.

This patch was tested by simulating the update that goes from 19460 to
19580 with the desktop pack installed (that contains one hardlink in
the pack used by this update).

Fixes #351.