Remove root requirement for check-update
Also removes a call to update an unused motd file. The motd is now
managed differently in Clear Linux and the MOTD_FILE does not even exist
where it expects it to. Since this was the only reason root was required
just remove the functionality entirely.

Fixes #398 (the best we can do for now)

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>