update: allow update to a specific version using -m
Allow updating to a version greater than current using -m. Also
supports 'latest' that has the same behavior as not passing any flag.

If the requested version is smaller than the server version (which
is the latest one available in the current format), use that. If not,
use the latest available and re-execute.

Care was taken so that a "update -m LAST_OF_A_BUMP" will not trigger a
re-execution of update.

At the moment version_url is still used, but a future improvement
could avoid it by first looking at the format of the given version
first.

Fixes #257.