update: do not retry when server manifest not found
Change load_mom to have a variant that can propagate the error code,
and check for that. Use that variant when downloading the
server_manifest. This approach could later be applied to other retries
that involve downloading.

With this patch, "update -m" to a non-published version will not retry
many times before failing.