Add test to ensure man pages are up to date
Current idea is to have the manual pages stored in the repo in nroff
format, as well as restructured test format. Users can just use the
nroff. Add test to ensure that the pages match.

Fixes #316
Fixes #425

Make sure that running "make" hasn't fixed the local manual pages but
not checked them in.

Signed-off-by: Icarus Sparry <icarus.w.sparry@intel.com>