Avoid retries in local downloads for manifests and packs
If content_url is local, do not attempt certain certain download
retries. I've focused on changing only the spots that caused 'make
check' to sleep as a first step. But I agree the retry logic needs
some extra care as posted in #221.

This patch reduces the time of running 'make check' in my system from
2m44s to 24s.