Release v3.17.0
This release adds configurable mirror support to swupd via the 'mirror'
subcommand. Users can 'set' or 'unset' a mirror and the relevant
configuration files will be placed in /etc/swupd/. If the set mirror is
detected to be a small amount behind the upstream CDN a warning will be
issued to the user. If the mirror is significantly out-of-date the
mirror will be automatically unset to supply the user the most recent
security updates.

SSL certificate failures are now handled specially to provide meaningful
telemetry and error reporting to the user.

Retries are avoided when the contenturl is a file:// url to a local
update directory.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>