mixer integration: check for file collisions after new build
Fixes #446
When a new upstream version is detected the new local mixer build must
be done before manifest uniqueness is checked. This is due to the fact
that os-core can change in the upstream version, so it has to be
re-built locally against the upstream content in order to avoid false
positives for file colisions.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>