Remove all commented code from the code base
We use a full feature version control system (git) for a reason.

'non-production' code can be kept on each one favorite personal branch
and debug code, if key for development, should be implemented as
working/tested part of the project (via a well designed logging system)
and not as commented code.

Not to mention that this is be a best pratice which we should all watch
over.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>