curl: Reduce coupling of curl module by removing information about version
Curl module doesn't need to have any information on what is the content of
the file it's going to download. And there's no reason to force users to
use memory to store a file only for version numbers. If there's another
small file needed in the future, we can use the memory to hold it.

Also fixes:
 - Memory allocated for version string was too big. Using maximum number of
   chars to store an integer for it.
 - As the size to store a version string is small, uses the stack instead of
   the heap for it.
 - On swupd_curl_test_resume, don't alocate memory and copy data to it
   to be discarded later.