Verify --install: Download full files if pack download fails
Use fullfile_download to download full files from server if a pack
download has failed or if the file is missing/corrupted in a pack.

Verify --install wasn't checking if the content of the downloaded
files match the hash in the manifest. As the fullfile download
function check that, the hashes in some tests needed to be fixed.