Optimize bundle-add to only install files that don't yet exist
With overlapping bundle-content, it's not entirely uncommon for files
that are in a newly added bundle to already exist on the system.

This patch optimizes this case by removing these from the files that need
to be installed.

In addition, if after this filter step, there's less than 5
(number to be tuned) files to install left, skip downloading the pack
for the bundle and just grab the full files directly.