Release v3.17.5
This release adds a file to the ignore list for hash comparisons during
migrating to a local mix due to that file being modified by the mix
itself (in an expected manner). This release also adds an up-to 10x
improvement for bundle-adds of small bundles and bundles where most of
the content is already on the system.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>