verify: Ignore mix state when fixing to version
When the version the user is fixing to is different from the version
they are on disable checks for mixer state on the system. This allows
users on local mixes to easily migrate back to the upstream update
stream. It also means users are not able to fix to another version
within their own mix.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>