download: remove mutex lock in single thread code
Functions clean_curl_multi_queue() and swupd_curl_hashmap_insert() are
always called in the same thread so there isn't any need to protect
access to swupd_curl_hashbucket or swupd_curl_hashmap.

The download.c functions use curl_multi infrastructure to download in parallel
but all processing is performed in the main thread. So we don't need to worry
about concurrency problems in object access.

But we need to worry about concurrency in file access. Fortunatelly the hashmap
prevents the same file to be downloaded twice, so we are safe on that.