download: Make parallel downloads generic
Remove all information about the content of the file that the parallel download
element is downloading. Move everything related to fullfile downloads to
fullfile.c and make the parallel downloader ready to be used for other
download cases.

There were some changes in this commit that were really hard to split
in individual commits, like:
 - Create a struct curl_file for curl specific information instead of
   adding extra fields to a generic file struct
 - Rename swupd_download_file_start() to swupd_download_file_create(),
   because that's what this function is doing.