Rewriting bundle-list tests with the new testlib
A new test library.bash was introduced that simplifies the creation
of test environments and test objects.

This commit rewrites all existing tests from test/functional/bundlelist
but this time using the new test library. Once all the tests have been
rewriten to use the new lib, they can be included in the Makefile so
they are run with every patch, and the  old tests can be deleted.