Release v3.17.9
This release makes several internal improvements to the swupd codebase,
adds a few bugfixes, and adds pre-work for future features. An update to
the swupd manpage was added, unnecessary bundle untracking code was
removed, unused and unnecessary version number reservations were
removed, a policy warning was added to autoupdate disabling, an output
message bug in bundle-add was fixed, and over-reporting on file
descriptor leaks was fixed as well.

Swupd will now track which bundles are manually installed on the system
vs pulled in by bundle includes. This is pre-work for a potential future
feature for recursive removal of included bundles when doing a
bundle-remove.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>