Add remove and install bundle functions to testlib
This commit adds the following three funtions to the test library:
- install_bundle
- remove_bundle
- clean_state_dir

These functions are useful specially when the global_setup and the
global_teardown are used instead of test_setup and test_teardown.

These functions allow the user to create all of its test resources
in the global_setup and reuse those same resources throughout all
the tests in the test script. By using the global_setup and
global_teardown functions the test execution time can be greatelly
reduced.

This commit also modifies the add-rc.bats and remove-rc.bats to
leverage the global setup and teardowns to increase the execution
speed. These times were reduced as shown:
 - bundleadd_v2/add-rc.bats from 1m50.129s -> 0m20.026s
 - bundleremove_v2/remove-rc.bats from 0m59.079s -> 0m15.446s