manifest: ignore I flag in manifest
The 'I' flag will be used to represent iterative manifests once the
feature becomes available. For now we need to ignore this flag so the
feature can be implemented in mixer before implementation here. Before
this patch manifest parsing would fail if it encountered an 'I' flag.
The iterative manifests will be published and included in the
Manifest.MoM before actually being used by client, which is why it is
important to ignore and not fail here.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>